#50 move textEditor module to ui package
