+++ ui/feature-home/build.gradle.kts
@@ -20,7 +20,7 @@ kotlin {
         desktopMain.dependencies {
             implementation(compose.desktop.currentOs)
             implementation(libs.kotlinx.coroutines.swing)
-            implementation(projects.texteditor)
+            implementation(projects.ui.texteditor)
         }
     }
 }
