+++ ui/snackbar/build.gradle.kts
@@ -7,7 +7,7 @@ plugins {
 }
 
 android {
-    namespace = "${Configuration.PACKAGE_NAME}.home"
+    namespace = "${Configuration.PACKAGE_NAME}.ui.snackbar"
 }
 
 kotlin {
