+++ ui/texteditor/.gitignore
