+++ ui/texteditor/proguard-rules.pro
