+++ ui/texteditor/src/commonMain/kotlin/com/haghpanah/pienote/utils/Common.kt
