+++ ui/texteditor/src/desktopMain/kotlin/com/haghpanah/pienote/shortcuthandler/KeyboardShortcutsManager.kt
