+++ ui/texteditor/src/desktopMain/kotlin/com/haghpanah/pienote/shortcuthandler/ShortcutType.kt
